Rename gcs and abs connectors
