+++ connectors/sources/azure_blob_storage.py
@@ -26,7 +26,9 @@ BLOB_SCHEMA = {
 DEFAULT_CONTENT_EXTRACTION = True
 DEFAULT_FILE_SIZE_LIMIT = 10485760
 DEFAULT_RETRY_COUNT = 3
-MAX_CONCURRENT_DOWNLOADS = 100  # Max concurrent download supported by abs
+MAX_CONCURRENT_DOWNLOADS = (
+    100  # Max concurrent download supported by Azure Blob Storage
+)
 
 
 class AzureBlobStorageDataSource(BaseDataSource):
