+++ connectors/sources/tests/fixtures/google_cloud_storage/.env
