+++ connectors/sources/tests/fixtures/google_cloud_storage/fixture.py
@@ -3,7 +3,7 @@
 # or more contributor license agreements. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with the Elastic License 2.0.
 #
-"""Google Cloud Storage module responsible to generate blob(s) on the fake gcs server.
+"""Google Cloud Storage module responsible to generate blob(s) on the fake Google Cloud Storage server.
 """
 import os
 import random
@@ -27,7 +27,7 @@ class PrerequisiteException(Exception):
 
     def __init__(self, errors):
         super().__init__(
-            f"Error while running e2e test for the GCS connector. \nReason: {errors}"
+            f"Error while running e2e test for the Google Cloud Storage connector. \nReason: {errors}"
         )
         self.errors = errors
 
@@ -44,7 +44,7 @@ def verify():
 
 
 def create_connection():
-    """Method for creating connection to the fake gcs server"""
+    """Method for creating connection to the fake Google Cloud Storage server"""
     try:
         global client_connection
         client_connection = storage.Client(
@@ -53,30 +53,30 @@ def create_connection():
         )
     except Exception as error:
         print(
-            f"Error occurred while creating connection to the fake gcs server. Error: {error}"
+            f"Error occurred while creating connection to the fake Google Cloud Storage server. Error: {error}"
         )
         raise
 
 
 def create_buckets():
-    """Method for generating buckets on the fake gcs server"""
+    """Method for generating buckets on the fake Google Cloud Storage server"""
     try:
-        print("Started loading 2 buckets on the fake gcs server....")
+        print("Started loading 2 buckets on the fake Google Cloud Storage server....")
 
         client_connection.create_bucket("sample-small-files-bucket")
         client_connection.create_bucket("sample-large-files-bucket")
 
-        print("Loaded total 2 buckets on the fake gcs server....")
+        print("Loaded total 2 buckets on the fake Google Cloud Storage server....")
     except Exception as error:
         print(
-            f"Error occurred while creating buckets on the fake gcs server. Error: {error}"
+            f"Error occurred while creating buckets on the fake Google Cloud Storage server. Error: {error}"
         )
 
 
 def generate_small_files(start_number_of_small_files, end_number_of_small_files):
-    """Method for generating small files on the fake gcs server"""
+    """Method for generating small files on the fake Google Cloud Storage server"""
     try:
-        print("Started loading small files on the fake gcs server....")
+        print("Started loading small files on the fake Google Cloud Storage server....")
         bucket = client_connection.bucket("sample-small-files-bucket")
 
         for number in range(start_number_of_small_files, end_number_of_small_files):
@@ -84,24 +84,24 @@ def generate_small_files(start_number_of_small_files, end_number_of_small_files)
             blob.upload_from_string("")
 
         print(
-            f"Loaded {end_number_of_small_files-start_number_of_small_files} small files on the fake gcs server...."
+            f"Loaded {end_number_of_small_files-start_number_of_small_files} small files on the fake Google Cloud Storage server...."
         )
     except Exception as error:
         print(
-            f"Error occurred while generating small files on the fake gcs server. Error: {error}"
+            f"Error occurred while generating small files on the fake Google Cloud Storage server. Error: {error}"
         )
         raise
 
 
 def generate_large_files(start_number_of_large_files, end_number_of_large_files):
-    """Method for generating large files on the fake gcs server"""
+    """Method for generating large files on the fake Google Cloud Storage server"""
     try:
         size_of_file = 2097152  # 2 Mb of text
         large_data = "".join(
             [random.choice(string.ascii_letters) for i in range(size_of_file)]
         )
 
-        print("Started loading large files on the fake gcs server....")
+        print("Started loading large files on the fake Google Cloud Storage server....")
 
         bucket = client_connection.bucket("sample-large-files-bucket")
 
@@ -110,11 +110,11 @@ def generate_large_files(start_number_of_large_files, end_number_of_large_files)
             blob.upload_from_string(large_data)
 
         print(
-            f"Loaded {end_number_of_large_files-start_number_of_large_files} large files on the fake gcs server...."
+            f"Loaded {end_number_of_large_files-start_number_of_large_files} large files on the fake Google Cloud Storage server...."
         )
     except Exception as error:
         print(
-            f"Error occurred while generating large files on the fake gcs server. Error: {error}"
+            f"Error occurred while generating large files on the fake Google Cloud Storage server. Error: {error}"
         )
         raise
 
@@ -127,15 +127,17 @@ def load():
     if NUMBER_OF_SMALL_FILES:
         generate_small_files(0, NUMBER_OF_SMALL_FILES)
     print(
-        f"Loaded 2 buckets with {NUMBER_OF_LARGE_FILES} large files and {NUMBER_OF_SMALL_FILES} small files on the fake gcs server."
+        f"Loaded 2 buckets with {NUMBER_OF_LARGE_FILES} large files and {NUMBER_OF_SMALL_FILES} small files on the fake Google Cloud Storage server."
     )
 
 
 def remove():
-    """Method for removing random blobs from the fake gcs server"""
+    """Method for removing random blobs from the fake Google Cloud Storage server"""
     create_connection()
     try:
-        print("Started removing random blobs from the fake gcs server....")
+        print(
+            "Started removing random blobs from the fake Google Cloud Storage server...."
+        )
 
         bucket = client_connection.bucket("sample-small-files-bucket")
         for number in range(0, NUMBER_OF_BLOBS_TO_BE_DELETED):
@@ -143,11 +145,11 @@ def remove():
             blob.delete()
 
         print(
-            f"Removed total {NUMBER_OF_BLOBS_TO_BE_DELETED} random blobs from the fake gcs server...."
+            f"Removed total {NUMBER_OF_BLOBS_TO_BE_DELETED} random blobs from the fake Google Cloud Storage server...."
         )
     except Exception as error:
         print(
-            f"Error occurred while removing blobs from the fake gcs server. Error: {error}"
+            f"Error occurred while removing blobs from the fake Google Cloud Storage server. Error: {error}"
         )
         raise
