+++ connectors/sources/tests/fixtures/google_cloud_storage/gcs_dummy_cert.pem
