+++ connectors/sources/tests/fixtures/google_cloud_storage/mocked_http_responses.json
