+++ connectors/sources/tests/fixtures/google_cloud_storage/mockerserver.properties
@@ -29,4 +29,4 @@ mockserver.enableCORSForAllResponses=true
 
 # Json Initialization
 
-mockserver.initializationJsonPath=/gcs/mocked_http_responses.json
+mockserver.initializationJsonPath=/google_cloud_storage/mocked_http_responses.json
