+++ connectors/sources/tests/fixtures/google_cloud_storage/service_account_dummy_cert.pem
